refactor: split block size tests
