+++ tests/block_size/delta_block_size.rs
@@ -1,89 +1,13 @@
-// tests/block_size.rs
-#![allow(clippy::needless_range_loop)]
-
 use assert_cmd::Command;
 use checksums::ChecksumConfigBuilder;
 use compress::available_codecs;
-use engine::{Op, SyncOptions, block_size, compute_delta, sync};
+use engine::{Op, SyncOptions, compute_delta, sync};
 use filters::Matcher;
 use std::fs;
 use tempfile::tempdir;
-mod common;
-use common::parse_literal;
+use super::common::parse_literal;
 
-#[test]
-fn block_size_matches_upstream() {
-    let data = fs::read_to_string("tests/golden/block_size/upstream_block_sizes.txt").unwrap();
-    for line in data.lines() {
-        let line = line.trim();
-        if line.is_empty() || line.starts_with('#') {
-            continue;
-        }
-        let mut parts = line.split_whitespace();
-        let len: u64 = parts.next().unwrap().parse().unwrap();
-        let expected: usize = parts.next().unwrap().parse().unwrap();
-        assert_eq!(block_size(len), expected, "len={len}");
-    }
-}
-
-#[test]
-fn delta_block_size_matches_rsync() {
-    for &block_size in &[1024usize, 2048usize, 4096usize] {
-        let dir = tempdir().unwrap();
-        let src_dir = dir.path().join("src");
-        let dst_dir = dir.path().join("dst");
-        fs::create_dir_all(&src_dir).unwrap();
-        fs::create_dir_all(&dst_dir).unwrap();
-        let src_file = src_dir.join("file.bin");
-        let dst_file = dst_dir.join("file.bin");
-
-        let size = 1 << 20;
-        let mut basis = vec![0u8; size];
-        for i in 0..size {
-            basis[i] = (i % 256) as u8;
-        }
-        let mut target = basis.clone();
-        let off = size / 2;
-        target[off..off + block_size].fill(0xFF);
-        fs::write(&src_file, &target).unwrap();
-        fs::write(&dst_file, &basis).unwrap();
-
-        let cfg = ChecksumConfigBuilder::new().build();
-        let mut basis_f = fs::File::open(&dst_file).unwrap();
-        let mut target_f = fs::File::open(&src_file).unwrap();
-        let ops: Vec<Op> = compute_delta(
-            &cfg,
-            &mut basis_f,
-            &mut target_f,
-            block_size,
-            usize::MAX,
-            &SyncOptions::default(),
-        )
-        .unwrap()
-        .map(Result::unwrap)
-        .collect();
-        let literal: usize = ops
-            .iter()
-            .map(|op| match op {
-                Op::Data(d) => d.len(),
-                _ => 0,
-            })
-            .sum();
-
-        let stats = fs::read_to_string(format!(
-            "tests/golden/block_size/delta_block_size_matches_rsync_bs{block_size}.stdout"
-        ))
-        .unwrap();
-        let rsync_literal = parse_literal(&stats);
-        assert_eq!(literal, rsync_literal);
-        assert_eq!(literal, block_size);
-    }
-}
-
-#[test]
-fn delta_block_size_large_file() {
-    let block_size = 8192usize;
-    let size = 8 * 1024 * 1024;
+fn run_delta_block_size(block_size: usize, golden: &str) {
     let dir = tempdir().unwrap();
     let src_dir = dir.path().join("src");
     let dst_dir = dir.path().join("dst");
@@ -92,13 +16,14 @@ fn delta_block_size_large_file() {
     let src_file = src_dir.join("file.bin");
     let dst_file = dst_dir.join("file.bin");
 
+    let size = 1 << 20;
     let mut basis = vec![0u8; size];
     for i in 0..size {
         basis[i] = (i % 256) as u8;
     }
     let mut target = basis.clone();
     let off = size / 2;
-    target[off..off + block_size].fill(0xAA);
+    target[off..off + block_size].fill(0xFF);
     fs::write(&src_file, &target).unwrap();
     fs::write(&dst_file, &basis).unwrap();
 
@@ -124,11 +49,34 @@ fn delta_block_size_large_file() {
         })
         .sum();
 
-    let stats =
-        fs::read_to_string("tests/golden/block_size/delta_block_size_large_file.stdout").unwrap();
+    let stats = fs::read_to_string(golden).unwrap();
     let rsync_literal = parse_literal(&stats);
     assert_eq!(literal, rsync_literal);
     assert_eq!(literal, block_size);
+}
+
+#[test]
+fn delta_block_size_matches_rsync_1024() {
+    run_delta_block_size(
+        1024,
+        "tests/golden/block_size/delta_block_size_matches_rsync_bs1024.stdout",
+    );
+}
+
+#[test]
+fn delta_block_size_matches_rsync_2048() {
+    run_delta_block_size(
+        2048,
+        "tests/golden/block_size/delta_block_size_matches_rsync_bs2048.stdout",
+    );
+}
+
+#[test]
+fn delta_block_size_matches_rsync_4096() {
+    run_delta_block_size(
+        4096,
+        "tests/golden/block_size/delta_block_size_matches_rsync_bs4096.stdout",
+    );
 }
 
 #[test]
@@ -176,8 +124,7 @@ fn delta_block_size_unaligned_edit() {
         .sum();
 
     let stats =
-        fs::read_to_string("tests/golden/block_size/delta_block_size_unaligned_edit.stdout")
-            .unwrap();
+        fs::read_to_string("tests/golden/block_size/delta_block_size_unaligned_edit.stdout").unwrap();
     let rsync_literal = parse_literal(&stats);
     assert_eq!(literal, rsync_literal);
     assert_eq!(literal, block_size * 2);
@@ -227,8 +174,8 @@ fn delta_block_size_non_power_two() {
         })
         .sum();
 
-    let stats = fs::read_to_string("tests/golden/block_size/delta_block_size_non_power_two.stdout")
-        .unwrap();
+    let stats =
+        fs::read_to_string("tests/golden/block_size/delta_block_size_non_power_two.stdout").unwrap();
     let rsync_literal = parse_literal(&stats);
     assert!(literal >= rsync_literal);
 }
@@ -325,47 +272,6 @@ fn sync_block_size_literal_matches_rsync() {
 }
 
 #[test]
-fn sync_block_size_large_file_stats_match_rsync() {
-    let block_size = 8192usize;
-    let size = 8 * 1024 * 1024;
-    let dir = tempdir().unwrap();
-    let src_dir = dir.path().join("src");
-    let dst_dir = dir.path().join("dst");
-    fs::create_dir_all(&src_dir).unwrap();
-    fs::create_dir_all(&dst_dir).unwrap();
-    let src_file = src_dir.join("file.bin");
-    let dst_file = dst_dir.join("file.bin");
-
-    let mut basis = vec![0u8; size];
-    for i in 0..size {
-        basis[i] = (i % 256) as u8;
-    }
-    let mut target = basis.clone();
-    let off = size / 2;
-    target[off..off + block_size].fill(0xAA);
-    fs::write(&src_file, &target).unwrap();
-    fs::write(&dst_file, &basis).unwrap();
-
-    let stats = sync(
-        &src_dir,
-        &dst_dir,
-        &Matcher::default(),
-        &available_codecs(),
-        &SyncOptions {
-            block_size,
-            checksum: true,
-            ..Default::default()
-        },
-    )
-    .unwrap();
-    let rsync_stats =
-        fs::read_to_string("tests/golden/block_size/delta_block_size_large_file.stdout").unwrap();
-    let rsync_literal = parse_literal(&rsync_stats) as u64;
-    assert_eq!(stats.literal_data, rsync_literal);
-    assert_eq!(stats.literal_data, block_size as u64);
-}
-
-#[test]
 fn sync_block_size_unaligned_edit_stats_match_rsync() {
     let block_size = 1024usize;
     let size = 1 << 20;
@@ -400,8 +306,7 @@ fn sync_block_size_unaligned_edit_stats_match_rsync() {
     )
     .unwrap();
     let rsync_stats =
-        fs::read_to_string("tests/golden/block_size/delta_block_size_unaligned_edit.stdout")
-            .unwrap();
+        fs::read_to_string("tests/golden/block_size/delta_block_size_unaligned_edit.stdout").unwrap();
     let rsync_literal = parse_literal(&rsync_stats) as u64;
     assert_eq!(stats.literal_data, rsync_literal);
     assert_eq!(stats.literal_data, (block_size * 2) as u64);
@@ -475,21 +380,19 @@ fn cli_block_size_matches_rsync() {
     );
 }
 
-#[test]
-fn cli_block_size_errors_match_rsync() {
+fn run_cli_error(prefix: &str, args: &[&str]) {
     let tmp = tempdir().unwrap();
     let dst = tmp.path().join("dst");
     fs::create_dir_all(&dst).unwrap();
 
     let expected_code: i32 =
-        fs::read_to_string("tests/golden/block_size/cli_block_size_errors_match_rsync.exit")
+        fs::read_to_string(format!("tests/golden/block_size/{}.exit", prefix))
             .unwrap()
             .trim()
             .parse()
             .unwrap();
     let rsync_err =
-        fs::read_to_string("tests/golden/block_size/cli_block_size_errors_match_rsync.stderr")
-            .unwrap();
+        fs::read_to_string(format!("tests/golden/block_size/{}.stderr", prefix)).unwrap();
     fn sanitize(line: &str) -> &str {
         line.split_once(' ').map(|(_, rhs)| rhs).unwrap_or(line)
     }
@@ -501,70 +404,51 @@ fn cli_block_size_errors_match_rsync() {
         .unwrap_or(rsync_lines[1]);
     let rsync_second_prefix = sanitize(rsync_second_lhs);
 
-    for args in [&["--block-size=1x"][..], &["-B1x"][..], &["-B", "1x"][..]] {
-        let ours = Command::cargo_bin("oc-rsync")
-            .unwrap()
-            .args(args)
-            .arg("/dev/null")
-            .arg("--")
-            .arg(dst.to_str().unwrap())
-            .output()
-            .unwrap();
-        assert_eq!(ours.status.code(), Some(expected_code));
-        let ours_err = String::from_utf8(ours.stderr).unwrap();
-        let ours_lines: Vec<_> = ours_err.lines().collect();
-        assert_eq!(rsync_first, sanitize(ours_lines[0]));
-        let ours_second_lhs = ours_lines[1]
-            .split_once(" at ")
-            .map(|(lhs, _)| lhs)
-            .unwrap_or(ours_lines[1]);
-        assert_eq!(rsync_second_prefix, sanitize(ours_second_lhs));
-    }
+    let ours = Command::cargo_bin("oc-rsync")
+        .unwrap()
+        .args(args)
+        .arg("/dev/null")
+        .arg("--")
+        .arg(dst.to_str().unwrap())
+        .output()
+        .unwrap();
+    assert_eq!(ours.status.code(), Some(expected_code));
+    let ours_err = String::from_utf8(ours.stderr).unwrap();
+    let ours_lines: Vec<_> = ours_err.lines().collect();
+    assert_eq!(rsync_first, sanitize(ours_lines[0]));
+    let ours_second_lhs = ours_lines[1]
+        .split_once(" at ")
+        .map(|(lhs, _)| lhs)
+        .unwrap_or(ours_lines[1]);
+    assert_eq!(rsync_second_prefix, sanitize(ours_second_lhs));
 }
 
 #[test]
-fn cli_block_size_zero_errors_match_rsync() {
-    let tmp = tempdir().unwrap();
-    let dst = tmp.path().join("dst");
-    fs::create_dir_all(&dst).unwrap();
+fn cli_block_size_errors_long_form() {
+    run_cli_error("cli_block_size_errors_match_rsync", &["--block-size=1x"]);
+}
 
-    let expected_code: i32 =
-        fs::read_to_string("tests/golden/block_size/cli_block_size_zero_errors_match_rsync.exit")
-            .unwrap()
-            .trim()
-            .parse()
-            .unwrap();
-    let rsync_err =
-        fs::read_to_string("tests/golden/block_size/cli_block_size_zero_errors_match_rsync.stderr")
-            .unwrap();
-    fn sanitize(line: &str) -> &str {
-        line.split_once(' ').map(|(_, rhs)| rhs).unwrap_or(line)
-    }
-    let rsync_lines: Vec<_> = rsync_err.lines().collect();
-    let rsync_first = sanitize(rsync_lines[0]);
-    let rsync_second_lhs = rsync_lines[1]
-        .split_once(" at ")
-        .map(|(lhs, _)| lhs)
-        .unwrap_or(rsync_lines[1]);
-    let rsync_second_prefix = sanitize(rsync_second_lhs);
+#[test]
+fn cli_block_size_errors_short_combined() {
+    run_cli_error("cli_block_size_errors_match_rsync", &["-B1x"]);
+}
 
-    for args in [&["--block-size=0"][..], &["-B0"][..], &["-B", "0"][..]] {
-        let ours = Command::cargo_bin("oc-rsync")
-            .unwrap()
-            .args(args)
-            .arg("/dev/null")
-            .arg("--")
-            .arg(dst.to_str().unwrap())
-            .output()
-            .unwrap();
-        assert_eq!(ours.status.code(), Some(expected_code));
-        let ours_err = String::from_utf8(ours.stderr).unwrap();
-        let ours_lines: Vec<_> = ours_err.lines().collect();
-        assert_eq!(rsync_first, sanitize(ours_lines[0]));
-        let ours_second_lhs = ours_lines[1]
-            .split_once(" at ")
-            .map(|(lhs, _)| lhs)
-            .unwrap_or(ours_lines[1]);
-        assert_eq!(rsync_second_prefix, sanitize(ours_second_lhs));
-    }
+#[test]
+fn cli_block_size_errors_short_separate() {
+    run_cli_error("cli_block_size_errors_match_rsync", &["-B", "1x"]);
+}
+
+#[test]
+fn cli_block_size_zero_errors_long_form() {
+    run_cli_error("cli_block_size_zero_errors_match_rsync", &["--block-size=0"]);
+}
+
+#[test]
+fn cli_block_size_zero_errors_short_combined() {
+    run_cli_error("cli_block_size_zero_errors_match_rsync", &["-B0"]);
+}
+
+#[test]
+fn cli_block_size_zero_errors_short_separate() {
+    run_cli_error("cli_block_size_zero_errors_match_rsync", &["-B", "0"]);
 }
