+++ tests/block_size/mod.rs
@@ -1,0 +1,6 @@
+#[path = "../common/mod.rs"]
+mod common;
+
+mod unit_block_size;
+mod delta_block_size;
+mod large_file;
